feat: implement logic to filter ghosts based on observations
Add javascript code to handle selecting and de-selecting evidence in the
left column and have that propagate up to the `App` class.
Implement javascript code to filter the ghosts visible in the right
column based on the evidence selected in the left column. Selected evidence
will also show an icon next to it in the right column.
Add new classes to store Ghost and Evidence data.
Add metadata for initial set of ghosts and evidence.