feat: support for rejecting evidence
Add a new checkbox to the evidence selection to allow both selecting a
piece of evidence, or rejecting it (meaning that that evidence was not
observed).
Update the logic filtering which ghosts are possible based on the new
rejected evidence as well as the selected evidence.
Update the logic for missing evidence to include non-rejected evidence.