feat: show selected evidence in ghost modal
Add a new class to handle rendering ghost evidence tags. This new class
is used to render evidence tags in multiple places, including the ghost
info modal.
Move logic to determine if a tag was selected or not to the new evidence
tag to support showing which ghost evidence has been selected in the ghost
info modal.
Move ghost info modal icon to the right of the ghost name to follow a more
standard convention.