feat: show version number and build date on page (#1)
Add a new section in the left column that includes the version number and
the app build date. This data is loaded from environment variables.
Update the project makefile to set the necessary variables during the
build process.
Remove the version property from the project `package.json` file as it
serves no purpose and causes confusion.