feat: show matching ghost number for each evidence
Update the evidence toggles on the left column to show the number of ghosts
that include each piece of evidence along with the total number of possible
ghosts.
Change the style of the "select" button on evidence that has zero possible
ghosts to give a more visual indication that the evidence is not valid.