build: new review app for merge request preview
Update the Gitlab CI/CD pipeline to configure a new review app for each
build of the app.
Add a `PUBLIC_URL` variable to the build step to update the path where the
react app expects to be hosted. In this case, it will be hosted out of the
job artifacts folder.
Configure the build pipeline to only run the publish build on the main
branch, and run the preview build in all other cases.