refactor: add json files for data
Replace the hardcoded evidence and ghost data in the `index.js` file with
two new `.json` files that store all the necessary data. These files also
replace the class definitions for `Evidence` and `Ghost`.
This opens the possibility of the ghost and evidence data being generated
automatically in the future.