ci: use node 16 for build process
* Replace docker image used in CI/CD
Replace the node docker image used in the CI/CD process to use node 16
instead of the latest possible node version. This should minimize the
risk of builds failing because the build using a node version that is
incompatible with the existing source-code. In other words, builds failing
because the environment changed.