refactor: use functional components (#10)
* Refactor code to use functional components
Re-write all react components to use the function format instead of using
the older class-style component format. This allows the use of react hooks
for certain things like setting up state.

* Use a simple array for evidence
Refactor the `evidence.json` file to be a simple array of strings instead
of an array of objects. The evidence data array is derived from the simple
array. This helps decouple the evidence data itself (the name) from app
specific concepts such as `selected` and `rejected`.