feat: add extra evidence to mimic ghost
Update the evidence for the Mimic ghost to include Ghost Orb. Since this
evidence may or may not appear for this ghost, it essentially has 4 pieces
of evidence instead of 3. This change helps avoid selecting the wrong ghost
because Ghost Orb was seen instead of one of the ghost's real evidence.